requirements.txt
python-telegram-bot[job-queue]>=20.0
langchain>=0.1.0
langchain-openai
langchain-community
langchain-chroma
chromadb
pymongo
python-dotenv
pypdf
tiktoken
docx2txt
unstructured
openpyxl
pandas
